Surface Selection and Drum Pad Creation
Surface Selection: Candidate surfaces (tables and floors) are highlighted and user must place both hands on a surface to select it as the desired drum surface

Drum Pad: Created a prefab for the drum pad and associated scripts. Drum pad can be interacted with using custom finger hitboxes.

Main Controller is the top-level class for the application's flow. Starts surface selection on start up. Controls which mode you're in.

Imported starting audio files.