Better Fingertips, Code clean up, and new Assets
Fingertips are now controlled by their prefab which is attached as a child of the fingertip. They now have a physical representation and are scaled properly for hand scale and each finger.  Hitboxes are much more accurate to actual fingers now.

Codebase contains organizational regions and scripts had to be changed to accomodate Fingertip prefab.

Audio files need to be edited and have 0 lead in. This was tested with the two edited audio files and is better for user responsiveness.

Materials created for drum heads and fingertips. Not final.